Finnished Part 4 of Assignment 1
Implemented a move sequence using two heaps. Added methods that print a
win and loose message. Player moves randomly and does not tunnel.
Implemented monster movement to specification, reversed dijkstra's
algorithm for that to happen. Revised the git_neighbor method to ignore
the boarder.

+++ heap.h
@@ -5,20 +5,20 @@ typedef struct heap_node heap_node_t;
 
 struct heap_node {
     void *value;
-    uint16_t priority;
+    uint32_t priority;
 };
 
 typedef struct heap {
     heap_node_t *nodes;
-    uint16_t length;
-    uint16_t size;
+    uint32_t length;
+    uint32_t size;
 } heap_t;
 
 int heap_init(heap_t *h);
 int heap_delete(heap_t *h);
-int heap_add(heap_t *h, void *value, uint16_t priority);
-int heap_decrease_priority_vertex(heap_t *h, void *value, uint16_t priority);
-int heap_peek(heap_t *h, void **value, uint16_t *priority);
+int heap_add(heap_t *h, void *value, uint32_t priority);
+int heap_decrease_priority_vertex(heap_t *h, void *value, uint32_t priority);
+int heap_peek(heap_t *h, void **value, uint32_t *priority);
 int heap_extract_min(heap_t *h, void **value);
 int heap_size(heap_t *h);
